tests: Refactor a couple of methods for mapper tests
