+++ tests/Dfe.PlanTech.AzureFunctions.UnitTests/Mappers/BaseMapperTests.cs
@@ -1,5 +1,7 @@
 using System.Text.Json;
+using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
+using Dfe.PlanTech.Domain.Caching.Models;
 
 namespace Dfe.PlanTech.AzureFunctions;
 
@@ -10,4 +12,27 @@ public abstract class BaseMapperTests
     PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
     Converters = { new JsonStringEnumConverter(JsonNamingPolicy.CamelCase) },
   };
+
+  public static Dictionary<string, object?> WrapWithLocalisation(object? toWrap, string localisation = "en-US")
+  => new()
+  {
+    [localisation] = toWrap
+  };
+
+
+  protected CmsWebHookPayload CreatePayload(Dictionary<string, object?> fields, string entityId)
+  {
+    var asJson = JsonSerializer.Serialize(fields, JsonOptions);
+    var asJsonNode = JsonSerializer.Deserialize<Dictionary<string, JsonNode>>(asJson, JsonOptions);
+
+    var payload = new CmsWebHookPayload()
+    {
+      Sys = new CmsWebHookSystemDetails()
+      {
+        Id = entityId
+      },
+      Fields = asJsonNode!
+    };
+    return payload;
+  }
 }
